Extract odata response related functions from sbvr-utils
Change-type: patch
